GL renderer: Cache the result of the font resolution in the item rendering cache
For rendering itself as well as font_metrics() queries, store the result
in the item's rendering cache. This will be needed as the lookup is
about to become more expensive due to the need of calling OS functions
that do text analysis to determine a list of fallback fonts.

When text hasn't changed but we re-render items or re-do the layout, we
need that to be cached.