Revert logo in C++ docs
This reverts the Pipenv pinning of
f9ef21bffddeb8ee18a2fb2aa8dfff18caf57034 and thus disables the logo, as
subsequent changes in the pinned sphinx fork cause unrelated breakages
when building docs. Let's re-enable this when the fixes are in an
upstream release of sphinx and the rtd theme.