Fix shuffle and auto-play checkbox in slide puzzle
Text's behavior right now is to clip on a line basis, so with the given
font-size for these two elements, the glyphs aren't shown at all because
the ascent + descent that's use for the clip test is bigger than the
maximum height, despite the actual glyphs fitting.

For now, reduce the font-size slightly to make the glyphs fit.