C++ docs: more intro touch-ups
Don't mention the tutorial twice, and there's no need to line to the online docs.
The latter makes sense in the GitHub README, but not in the docs themselves.

Would be nice to fold this "intro" somehow right into the contents
though, without a sub-page. Or the intro becomes bigger with more
references.