Janitor: Fix clippy error about a loop that never loops
This code abuses a loop for flow control. Split this out into a separate
function instead, using a custom enum to transfer the information back.

This is necessary since we need to transfer ownership of the
mouse_input_state back to the caller.

This change should not cause a behavior change.