Fix interpreter ffi generated by cbindgen
Make sure to never export the Value enum. It might drag in all sorts of
unrelated types (as it happened in #311). We have ValueOpaque for that,
so exclude Value from the export config explicitly. However since we
expose the Value *pointers* for the struct iterators, at least provide a
forward declaration.