Janitor: Replace `'` with `´` where that makes sense
Replace strings to use `´` for elisions instead of `'`. Sorry, this
is a pet peeve of mine;-)

Remove many of the elisions altogether by spelling them out. That looks
IMHO more professional in the generated documentation (where many of
these strings might show up).

On the plus side this makes cspell a bit happier, since that recognizes
that the part after the `´` can be ignored, which it does not
consistently do for the part after `'`. This allows to remove one extra
entry in the word list.