Fix the spinner for the slide puzzle firefox
The layout of the html here is a bit different for the slize puzzle
than for other demo because it is using a display: grid.
As a result, placing the spinner in the grid would cause the canvas
to be resized when the spinner goes away, with firefox. Since we don't
get resize event when the canvas gets resized withtout the whole page
being resized, the scaling and coordinate were off.

So put the spinner outside of the grid