GL backend: Prospective fix for panic when trying to create window with AMD GL driver
The srgb workaround from commit a0e827e2004efac1987427ae507732410fcaa9a9
appears to work with the AMD GL driver, without which the GL context
creation works (as tested by the reporter).

So instead of panicing, fall back to the glutin default with possibly an
srgb framebuffer, when trying to create a GL context. If that also
fails, provide a better error message than "called unwrap on an err".

cc #481