Enable resource embedding C++ build coverage
In the CMake build on Linux, enable resource embedding.

This is a compromise clearly - it would be best to enable it on all, but
OTOH this slows down compilation. There's also no test for whether this
works at run-time, it would be best to have an API for that, i.e. the
ability to annotate an @image-url.