Minor cleanup in pass handling
Remove the roots vector again, with selective inlining it's better
to use the "live" data structures instead of a copy of the roots.

It would be nice to centralize that into a helper function that returns
say an `Ref<'_, impl Iterator<Item = &Rc<Component>>>>` but that's not
possible yet.

The same chained iterator pattern is also used in some passes already.