Minor cleanup in C++ code generator
In the tree builder's enter_component we registered the sub-component as a member
and also registered the member initializer
in the constructor.

However the call to `init()` to initialize the bindings was registered in `enter_component_children` for no good reason.

It's cleaner to do that all in one place.