Fix item tree offsets for chained sub-components
When we have

```
S1 := Rectangle {}
S2 := S1 {}
App := Window {
    S2 {}
}
```

Then the right path to the rectangle is

    offsetof(App, s2) + offsetof(S2, root) + offset(S1, rectangle)

The middle one was left out in C++, where it also probably doesn't matter
because the root is the first member (although it might not always remain that way).

In Rust the FieldOffset type generated won't allow for adding
together without the middle one.

Therefore visit_item() in the generator is changed to bring forward the sub-component
state when following the chain.